requirements.txt
asgiref==3.8.1
Django==4.2.18
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework_simplejwt==5.4.0
PyJWT==1.7.1
sqlparse==0.5.3
typing_extensions==4.12.2
